+++ src/main/java/com/hexvane/strangematter/item/TinfoilHatArmorMaterial.java
@@ -87,3 +87,4 @@ public enum TinfoilHatArmorMaterial implements ArmorMaterial {
         return this.knockbackResistance;
     }
 }
+